Fix invariant support
