+++ lib/src/main/kotlin/app/cash/kfsm/State.kt
@@ -4,28 +4,33 @@ package app.cash.kfsm
  * Base class for defining states in a finite state machine.
  *
  * States define the allowed transitions in the state graph. Each state knows
- * which states it can transition to directly and can define invariants that
- * must hold true while a value is in that state.
+ * which states it can transition to directly.
+ *
+ * To define invariants that must hold while a value is in a state, implement
+ * [StateWithInvariants] on your sealed state class.
  *
  * Example:
  * ```kotlin
- * sealed class OrderState : State<OrderState>() {
+ * sealed class OrderState : State<OrderState>(), StateWithInvariants<Order> {
+ *   // Default: no invariants
+ *   override fun invariants(): List<Invariant<Order>> = emptyList()
+ *
  *   data object Pending : OrderState() {
  *     override fun transitions() = setOf(Confirmed, Cancelled)
  *   }
  *   data object Confirmed : OrderState() {
  *     override fun transitions() = setOf(Shipped, Cancelled)
- *     override fun <V> invariants() = listOf(
- *       invariant("Order must have payment reference") { v: V ->
- *         (v as? Order)?.paymentReference != null
+ *     override fun invariants() = listOf(
+ *       invariant("Order must have payment reference") { order ->
+ *         order.paymentReference != null
  *       }
  *     )
  *   }
  *   data object Shipped : OrderState() {
  *     override fun transitions() = setOf(Delivered)
- *     override fun <V> invariants() = listOf(
- *       invariant("Shipped orders must have tracking number") { v: V ->
- *         (v as? Order)?.trackingNumber != null
+ *     override fun invariants() = listOf(
+ *       invariant("Shipped orders must have tracking number") { order ->
+ *         order.trackingNumber != null
  *       }
  *     )
  *   }
@@ -45,16 +50,6 @@ abstract class State<S : State<S>> {
    * Returns the set of states that can be reached directly from this state.
    */
   abstract fun transitions(): Set<S>
-
-  /**
-   * Returns the list of invariants that must hold true while a value is in this state.
-   *
-   * Override this method to define state-specific invariants. Invariants are validated
-   * after a transition completes, and if any invariant fails, the transition is rejected.
-   *
-   * @return List of invariants for this state (empty by default)
-   */
-  open fun <V> invariants(): List<Invariant<V>> = emptyList()
 
   /**
    * The set of states that can be reached directly from this state through a single transition.
@@ -95,22 +90,5 @@ abstract class State<S : State<S>> {
     }
 
     return visited
-  }
-
-  /**
-   * Validates that a value satisfies all invariants defined for this state.
-   *
-   * @param value The value to validate
-   * @return Success with Unit if all invariants pass, or failure with the first violation
-   */
-  fun <V> validateInvariants(value: V): Result<Unit> {
-    val stateInvariants: List<Invariant<V>> = invariants()
-    for (invariant in stateInvariants) {
-      val result = invariant.validate(value)
-      if (result.isFailure) {
-        return result
-      }
-    }
-    return Result.success(Unit)
   }
 }
